requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.35.0
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.7.1
python-dotenv==1.0.1
